Fuck fucking H2 Database, I'm not ready. This is back to normal

+++ Backend/src/main/java/com/Maxeeey/TODOListElements/ITODOListElement.java
@@ -3,6 +3,6 @@ package com.Maxeeey.TODOListElements;
 public interface ITODOListElement {
 	public String getName();
 	public boolean getIsDone();
-	public Long getId();
+	public int getId();
 	public void changeIsDoneBoolean();
 }
